requirements.txt
matplotlib==2.1.2
torch==0.3.0.post4
opencv_python==3.4.0.12
six==1.10.0
numpy==1.13.1
pandas==0.22.0
tqdm==4.19.5
setuptools==33.1.1.post20170320
torchvision==0.2.0
scipy==1.0.0
Cython==0.28.2
Pillow==5.1.0
caffe2==0.8.1
cityscapesscripts==1.0.4
protobuf==3.5.2.post1
h5py==2.7.1
mock==2.0.0
future==0.16.0
progressbar2==3.37.1
pycocotools==2.0.0
scikit_learn==0.19.1
PyYAML==3.12
